Project TG Alterando descrição dos comandos
