+++ Controllers/ControllerRace.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Controllers;
+
+class ControllerRace
+{
+
+}
