+++ actions/adicionarCarros
@@ -3,11 +3,9 @@
 
 require_once __DIR__ . "/../vendor/autoload.php";
 
-use Models\Car;
+use Controllers\ControllerCar;
 
 $result = $_SERVER['argv'];
-
-$car = new Car();
 
 for ($i = 1; $i <= 5; $i++) {
     if (!isset ($result[$i])) {
@@ -17,4 +15,5 @@ for ($i = 1; $i <= 5; $i++) {
     }
 }
 
+$car = new ControllerCar();
 $car->newCar($result[1], $result[2], $result[3], $result[4], $result[5]);
