+++ composer.json
@@ -4,7 +4,8 @@
       "Tests/"
     ],
     "psr-4": {
-      "Models\\" : "Models/"
+      "Models\\" : "Models/",
+      "Controllers\\" : "Controllers/"
     }
   },
   "require": {
